requirements.txt
flask
flask-jwt-extended
flask-cors
flask-sqlalchemy
pymysql
bcrypt
python-dotenv
google-generativeai
